Vercel fix on export of routes
