+++ src/lib/auth.ts
@@ -1,0 +1,92 @@
+import { AuthOptions } from "next-auth"
+import CredentialsProvider from "next-auth/providers/credentials"
+import { Pool } from "pg"
+import bcrypt from "bcrypt"
+
+// Extend NextAuth types to include id
+declare module "next-auth" {
+  interface User {
+    id: string
+  }
+  interface Session {
+    user: {
+      id: string
+      name?: string | null
+      email?: string | null
+      image?: string | null
+    }
+  }
+}
+
+// Create connection pool for NextAuth
+const pool = new Pool({
+  connectionString: process.env.DATABASE_URL,
+})
+
+export const authOptions: AuthOptions = {
+  providers: [
+    CredentialsProvider({
+      name: "credentials",
+      credentials: {
+        email: { label: "Email", type: "email" },
+        password: { label: "Password", type: "password" }
+      },
+      async authorize(credentials) {
+        if (!credentials?.email || !credentials?.password) {
+          return null
+        }
+
+        try {
+          // Query your existing User table
+          const result = await pool.query(
+            'SELECT * FROM "User" WHERE email = $1',
+            [credentials.email.toLowerCase().trim()]
+          )
+
+          if (result.rows.length === 0) {
+            return null
+          }
+
+          const user = result.rows[0]
+
+          // Verify password
+          const passwordMatch = await bcrypt.compare(credentials.password, user.password)
+
+          if (!passwordMatch) {
+            return null
+          }
+
+          // Return user object that matches NextAuth User type
+          return {
+            id: user.id,
+            email: user.email,
+            name: `${user.firstName} ${user.lastName}`,
+          }
+        } catch (error) {
+          console.error("Authentication error:", error)
+          return null
+        }
+      }
+    })
+  ],
+  session: {
+    strategy: "jwt",
+  },
+  pages: {
+    signIn: "/login",
+  },
+  callbacks: {
+    async jwt({ user, token }) {
+      if (user) {
+        token.uid = user.id
+      }
+      return token
+    },
+    async session({ session, token }) {
+      if (session?.user && token) {
+        session.user.id = token.uid as string
+      }
+      return session
+    },
+  },
+}
